feat(cli): add build subcommand
