+++ packages/@remax/cli/package.json
@@ -33,7 +33,6 @@
     "@types/ejs": "^2.6.1",
     "@types/mini-css-extract-plugin": "^0.2.0",
     "@types/node": "^10.12.18",
-    "@types/webpack-merge": "^4.1.3",
     "typescript": "^3.2.2"
   }
 }
