+++ packages/@remax/cli/src/index.ts
@@ -1,10 +1,15 @@
 import program from 'commander';
 
-import dev from './dev';
+import compile from './compile';
 
 program
   .command('dev')
   .description('start development')
-  .action(dev);
+  .action(compile({ watch: true }));
+
+program
+  .command('build')
+  .description('build the project')
+  .action(compile());
 
 program.parse(process.argv);
